+++ app/js/core/namespace.js
@@ -5,7 +5,7 @@ if (typeof console == 'undefined') { window.console = {log:function() {}}; }
 window.sofia = {};
 
 // version
-sofia.version = '3.5.0.beta-1';
+sofia.version = '3.5.2';
 
 // allows items to register themselves
 sofia.plugins = [];